17_AI_Avatar_Video_Sales/pyproject.toml
[project]
name = "17-ai-avatar-video-sales"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
